restructure table components
