+++ web-ui/src/components/tables/denied-table/denied-table.less
