+++ web-ui/src/components/tables/failures-table/failures-table.less
