+++ web-ui/src/components/tables/queue-table/queue-table.less
